refactor: rename test
